Undo splitting at NULL bytes
See issue #1. A lot of these NULL bytes are parts of the control
sequences.

This was obvious to start, but I assumed I would be able to filter
these extra strings if I cut them out.

Now, I think it may be a better plan to try and actually parse them into
useable attributes for text.